Move Quote inside projection actor
