+++ daemon/src/to_sse_event.rs
@@ -3,6 +3,7 @@ use crate::model::cfd::{
     Dlc, OrderId, Payout, Role, SettlementKind, UpdateCfdProposal, UpdateCfdProposals,
 };
 use crate::model::{Leverage, Position, TakerId, Timestamp, TradingPair};
+use crate::projection::{Price, Quote, Usd};
 use crate::{bitmex_price_feed, model};
 use bdk::bitcoin::{Amount, Network, SignedAmount, Txid};
 use rocket::request::FromParam;
@@ -12,64 +13,6 @@ use serde::{Deserialize, Serialize};
 use std::convert::TryInto;
 use time::OffsetDateTime;
 use tokio::sync::watch;
-
-#[derive(Debug, Clone)]
-pub struct Usd {
-    inner: model::Usd,
-}
-
-impl Usd {
-    fn new(usd: model::Usd) -> Self {
-        Self {
-            inner: model::Usd::new(usd.into_decimal().round_dp(2)),
-        }
-    }
-}
-
-impl From<model::Usd> for Usd {
-    fn from(usd: model::Usd) -> Self {
-        Self::new(usd)
-    }
-}
-
-impl Serialize for Usd {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        <Decimal as Serialize>::serialize(&self.inner.into_decimal(), serializer)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Price {
-    inner: model::Price,
-}
-
-impl Price {
-    fn new(price: model::Price) -> Self {
-        Self {
-            inner: model::Price::new(price.into_decimal().round_dp(2)).expect(
-                "rounding a valid price to 2 decimal places should still result in a valid price",
-            ),
-        }
-    }
-}
-
-impl From<model::Price> for Price {
-    fn from(price: model::Price) -> Self {
-        Self::new(price)
-    }
-}
-
-impl Serialize for Price {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        <Decimal as Serialize>::serialize(&self.inner.into_decimal(), serializer)
-    }
-}
 
 #[derive(Debug, Clone, Serialize)]
 pub struct Cfd {
@@ -250,12 +193,12 @@ pub struct CfdsWithAuxData {
 impl CfdsWithAuxData {
     pub fn new(
         rx_cfds: &watch::Receiver<Vec<model::cfd::Cfd>>,
-        rx_quote: &watch::Receiver<bitmex_price_feed::Quote>,
+        rx_quote: &watch::Receiver<Quote>,
         rx_updates: &watch::Receiver<UpdateCfdProposals>,
         role: Role,
         network: Network,
     ) -> Self {
-        let quote = rx_quote.borrow().clone();
+        let quote: bitmex_price_feed::Quote = rx_quote.borrow().clone().into();
         let current_price = match role {
             Role::Maker => quote.for_maker(),
             Role::Taker => quote.for_taker(),
@@ -487,21 +430,9 @@ fn to_tx_url_list(state: model::cfd::CfdState, network: Network) -> Vec<TxUrl> {
     }
 }
 
-#[derive(Debug, Clone, Serialize)]
-pub struct Quote {
-    bid: Price,
-    ask: Price,
-    last_updated_at: Timestamp,
-}
-
-impl ToSseEvent for bitmex_price_feed::Quote {
+impl ToSseEvent for Quote {
     fn to_sse_event(&self) -> Event {
-        let quote = Quote {
-            bid: self.bid.into(),
-            ask: self.ask.into(),
-            last_updated_at: self.timestamp,
-        };
-        Event::json(&quote).event("quote")
+        Event::json(self).event("quote")
     }
 }
 
@@ -537,9 +468,6 @@ fn available_actions(state: CfdState, role: Role) -> Vec<CfdAction> {
 mod tests {
     use super::*;
 
-    use rust_decimal_macros::dec;
-    use serde_test::{assert_ser_tokens, Token};
-
     #[test]
     fn state_snapshot_test() {
         // Make sure to update the UI after changing this test!
@@ -566,19 +494,5 @@ mod tests {
         assert_eq!(json, "\"Refunded\"");
         let json = serde_json::to_string(&CfdState::SetupFailed).unwrap();
         assert_eq!(json, "\"SetupFailed\"");
-    }
-
-    #[test]
-    fn usd_serializes_with_only_cents() {
-        let usd = Usd::new(model::Usd::new(dec!(1000.12345)));
-
-        assert_ser_tokens(&usd, &[Token::Str("1000.12")]);
-    }
-
-    #[test]
-    fn price_serializes_with_only_cents() {
-        let price = Price::new(model::Price::new(dec!(1000.12345)).unwrap());
-
-        assert_ser_tokens(&price, &[Token::Str("1000.12")]);
     }
 }
